+++ R/chronique.figure.depassementscalendaires.R
@@ -21,6 +21,7 @@
 #' @param format Défini le format d'enregistrement (par défaut .png)
 #' @keywords chronique
 #' @import glue
+#' @import scales
 #' @import tidyverse
 #' @export
 #' @examples